# remove unnecessary object allocation and regex to speedup str_to_sec