--- a/lib/rex/time.rb
+++ b/lib/rex/time.rb
@@ -37,25 +37,22 @@ module ExtTime
   # Converts a string in the form n years g days x hours y mins z secs.
   #
   def self.str_to_sec(str)
-    fields = str.split(/ /)
+    fields = str.split
     secs   = 0
 
     fields.each_with_index { |f, idx|
-      val = 0
       case f
-        when /^year/
-          val = 31536000
-        when /^day/
-          val = 86400
-        when /^hour/
-          val = 3600
-        when /^min/
-          val = 60
-        when /^sec/
-          val = 1
+      when /^year/
+        secs += 31536000 * fields[idx-1].to_i
+      when /^day/
+        secs += 86400 * fields[idx-1].to_i
+      when /^hour/
+        secs += 3600 * fields[idx-1].to_i
+      when /^min/
+        secs += 60 * fields[idx-1].to_i
+      when /^sec/
+        secs += 1 * fields[idx-1].to_i
       end
-
-      secs += val * fields[idx-1].to_i
     }
 
     secs