@busterb 2018-04-30
Actually, this code is not actually _used_ anywhere. I propose we delete ExtTime entirely. Someone got way overzealous with this.