@busterb 2018-04-30
This got added here: 1a73949451325f4607fdccea5e6dcf1c4f0ea995 , apparently by accident, and it was never used in the following 12 years as far as I can see.